refactor: extract header from layout
easier to think about
